Add files via upload
This website is based on tabs therefore most of the work is done in Admin.html, looking back at this project, I would not use a tab layout, rather I would create different pages for each section of the website. This website uses the same Google Firebase Database as  the Study Companion mobile application.